# src/modules/core/users/dto/create-student.dto.ts
import { IsInt, IsOptional, IsString, MaxLength, Min } from 'class-validator';
import { CreateUserDto } from './create-user.dto';

export class CreateStudentDto extends CreateUserDto {
  @IsString()
  parentUserId: string;

  @IsOptional()
  @IsInt()
  @Min(1)
  age?: number;

  @IsOptional()
  @IsString()
  @MaxLength(50)
  grade?: string;
}